certificates: update the UBI paths for certificates, address perms
Update the `templates/_certificates.tpl` to alter the mount and
paths for `/etc/pki/ca-trust/extracted`, removing `/pem` from the
end. This address concerns seen in some deployments of
`update-ca-trust` failing with permissions on a file witin the
`openssl/` subdirectory.

All of the content within `/etc/pki/ca-trust/extracted` is generated
on a per-run basis, so this is simply now inclusive all all content
generated by the scripted patterns of `update-ca-trust`.

Changelog: changed