gitaly: Migrate cgroups to `config` key
Migrate the subset of cgroups keys that are used directly by
Gitaly's config.toml into the new `config` section.

Remove the `cgroups.enabled` option in favour of setting a positive
integer for `cgroups.repositories.count`. This aligns with how cgroups
are enabeld in Omnibus and via Gitaly's `config.toml`.

Default values and placeholders are left as-is. Manual templating of the
keys is also left untouched. A subsequent commit in this series will
remove the manual templating and instead directly translate keys under
`config` into Gitaly's `config.toml`.