Enable cluster tests on security mirror MRs
We want to trigger cluster tests on the security mirror MRs when we are
working on security issues. We'll also have to extend our shared cluster
agents to the security mirror, but this will be done on:

https://gitlab.com/gitlab-org/distribution/infrastructure/kube-agents/-/merge_requests/25

After merging the above, we expect the .rule:skip_if_no_cluster to not
be a blocker for the security mirror anymore, so the jobs we'll start to
be triggered. This MR adds a rule (skip_in_security_if_not_mr_pipeline),
   to make sure that we don't accidentally enable the cluster test in
   all security mirror pipelines, otherwise we could be running
   duplicated pipelines that already ran on canonical. This will cause
   our mirror to on only hammer our CI test clusters, but also to cause
   test failures due to race conditions, as the mirrored commits would
   have the same helm release names.