Webservice: Default to IPv6 compatible binds
We disabled IPv6 compatible binds in https://gitlab.com/gitlab-org/charts/gitlab/-/merge_requests/4382
because of a issue with throttling of unauthenticated web traffic.

Now with https://gitlab.com/gitlab-org/gitlab/-/merge_requests/196981
rails should see the real client IP, even if the client or one of the
intermediate proxies was mapped to an IPv6 address, so we can default
to IPv6 compatible binds.

Closes https://gitlab.com/gitlab-org/charts/gitlab/-/issues/6084