feat: Run ClickHouse migrations in GitLab-Migrations chart conditionally
Related to https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/21436.

We are moving all migration workloads from the Deploy VM to Kubernetes for GitLab.com. There is no
support for running ClickHouse migrations using our public Helm charts at the moment. The
[documentation for ClickHouse migrations] recommends running `gitlab-rake` directly. We don't want
to run `gitlab-rake` directly on the Toolbox pods. I would like to explore whether we can integrate
ClickHouse migrations into the GitLab-Migrations chart.

We don't directly call `gitlab-rake` anywhere in our Helm chart. Instead, we create a script in this
repository, which will be included in the GitLab CNG[^1] image, and then call that script here. I
have added the script to CNG in a separate MR, which must be merged before this one. (More on this
later.)

ClickHouse is already configured in the ConfigMap for Webservice, Sidekiq, and Toolbox. I think that
including the configuration in the Migrations chart (at the very least) allows customers to at least
use the GitLab-Migrations chart by customizing the `migrations.scriptTpl` value.

[documentation for ClickHouse migrations]: https://docs.gitlab.com/integration/clickhouse/#run-clickhouse-migrations