feat: Enable automatic DB reindexing using CronJob
Related to https://gitlab.com/gitlab-com/gl-infra/delivery/-/work_items/21668

[Automatic database reindexing] is possible in an Omnibus installation by flipping some
configuration values inside the `gitlab.rb` file. As far as I can tell, Helm chart installations
that use CNG (Cloud Native GitLab) container images do not have a native method for runnig
reindexing periodically.

The Toolbox chart already has a CronJob for taking a backup. So, I am proposing that we add a
CronJob that will perform automatic database reindexing on a user-specified schedule.

The script that is mentioned in this MR is added to the Rails image in this CNG MR:
https://gitlab.com/gitlab-org/build/CNG/-/merge_requests/2712 As usual, before running the script,
we ensure that the dependencies (Postgres and Redis) are accessible.

The script will run a Rake task, which emulates the behavior from Omnibus, when the configuration
flags mentioned [here] are enabled:
https://gitlab.com/gitlab-org/omnibus-gitlab/-/blob/b26f58eb8a5e14be5ee3ddc6c16a8ecf9a462181/files/gitlab-cookbooks/gitlab/recipes/database_reindexing_enable.rb#L27

[Automatic database reindexing]: https://docs.gitlab.com/omnibus/settings/database/#automatic-database-reindexing
[here]: https://docs.gitlab.com/omnibus/settings/database/#automatic-database-reindexing

Changelog: added